Reorganize AppServer integration tests into top-level folder
