+++ AppServer/tsconfig.json
@@ -13,5 +13,5 @@
     "types": ["bun-types"],
     "skipLibCheck": true
   },
-  "include": ["src/**/*.ts"]
+  "include": ["src/**/*.ts", "integration/**/*.ts"]
 }
